feat #37 디자인 변경
