+++ backend/settings.gradle
@@ -1,1 +1,2 @@
 rootProject.name = 'devnear'
+
