+++ frontend/app/client/mypage/page.tsx
@@ -3,6 +3,7 @@
 import { useState, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 import api from '../../lib/axios';
+import { NotificationBell } from '@/components/notifications/NotificationProvider';
 import { notifyAuthChanged } from '../../lib/authEvents';
 import { Briefcase, User, Settings, LogOut, ChevronRight, Activity } from 'lucide-react';
 import { motion } from 'framer-motion';
@@ -122,13 +123,14 @@ export default function ClientMyPage() {
                     <span className="text-[#FF7D00]">Dev</span><span className="text-[#7A4FFF]">Near</span>
                 </div>
 
-                <div className="flex gap-6 items-center">
+                <div className="flex gap-4 items-center md:gap-6">
                     <button
                         onClick={() => router.push('/client/dashboard')}
                         className="text-xs font-bold text-zinc-500 hover:text-zinc-900 tracking-widest transition uppercase font-mono"
                     >
                         DASHBOARD
                     </button>
+                    <NotificationBell />
                     <button
                         onClick={() => alert("시스템 설계 중입니다. 다음 업데이트를 기다려주세요.")}
                         className="px-6 py-2.5 bg-[#FF7D00] text-white rounded-xl text-xs font-black tracking-widest hover:brightness-110 transition shadow-md shadow-orange-100 uppercase font-mono"
